Added a check to see whether to-be-merged participant DMs have identical column headers.
To apply this check, set acceptNonMatchingColumns to False. If column headers are not identical, an exception is raised.
To skip this check, set acceptNonMatchingColumns to True. In this case, the intersection of the two lists of column headers is used. The latter results in behaviour as before the current commit.


Signed-off-by: Lotje van der Linden <l.vanderlinden@cogsci.nl>